# core/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.UnsubscribeRequest:
    properties:
      entity_id:
        type: integer
      entity_type:
        type: string
      fcm_token:
        type: string
    type: object
...
  models.BFFResponse:
    properties:
      cached_at:
        type: string
      data: {}
      has_next:
        type: boolean
      has_prev:
        type: boolean
      source:
        description: '"cache", "upstream", "stale"'
        type: string
      success:
        type: boolean
      week_end:
        type: string
      week_start:
        type: string
    type: object
  models.Day:
    properties:
      day:
        type: string
      lessons:
        items:
          $ref: '#/definitions/models.Lesson'
        type: array
    type: object
...
  models.Lesson:
    properties:
      auditCorps:
        type: string
      faculty:
        type: string
      group:
        type: string
      id:
        type: integer
      lesson:
        type: string
      publishDate:
        type: string
      subgroupName:
        type: string
      teacher:
        type: string
      time:
        type: integer
      type_work:
        type: string
    type: object
...
  storage.ScheduleChange:
    properties:
      change_type:
        description: '"added", "removed", "modified"'
        type: string
      created_at:
        type: string
      entity_id:
        type: integer
      entity_type:
        type: string
      id:
        type: integer
      lesson_id:
        type: integer
      new_data:
        type: string
      old_data:
        type: string
    type: object
  storage.Subscription:
    properties:
      before_minutes:
        type: integer
      digest_time:
        type: string
      entity_id:
        type: integer
      entity_type:
        type: string
      fcm_token:
        type: string
      last_reminder_at:
        type: string
      notify_before_lesson:
        type: boolean
      notify_daily_digest:
        type: boolean
      notify_on_change:
        type: boolean
      subgroup:
        type: string
      timezone:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: High-performance BFF for Omsu university schedule.
  title: Omsu Schedule Mirror API
  version: "1.0"
paths:
...
  /changes/{type}/{id}:
    get:
      description: Returns recent schedule changes for a group, tutor, or auditory.
      parameters:
      - description: 'Entity type: group, tutor, auditory'
        in: path
        name: type
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.BFFResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/storage.ScheduleChange'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get schedule changes
      tags:
      - Changes
...
  /schedule/auditory/{id}:
    get:
      description: Returns the schedule for a specific auditory. Lazy-fetches if not
        in cache.
      parameters:
      - description: Auditory ID
        in: path
        name: id
        required: true
        type: integer
      - description: Week start date in YYYY-MM-DD format
        in: query
        name: week_start
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.BFFResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Day'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get auditory schedule
      tags:
      - Schedules
...
  /schedule/group/{id}:
    get:
      description: Returns the schedule for a specific group. Lazy-fetches if not
        in cache.
      parameters:
      - description: Group ID (real_group_id)
        in: path
        name: id
        required: true
        type: integer
      - description: Week start date in YYYY-MM-DD format
        in: query
        name: week_start
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.BFFResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Day'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get group schedule
      tags:
      - Schedules
...
  /schedule/tutor/{id}:
    get:
      description: Returns the schedule for a specific tutor. Lazy-fetches if not
        in cache.
      parameters:
      - description: Tutor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Week start date in YYYY-MM-DD format
        in: query
        name: week_start
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.BFFResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Day'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tutor schedule
      tags:
      - Schedules
...
  /subscribe:
    post:
      consumes:
      - application/json
      description: Creates or updates a notification subscription for a group, tutor,
        or auditory.
      parameters:
      - description: Subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/storage.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Subscribe to schedule notifications
      tags:
      - Notifications
...
  /unsubscribe:
    post:
      consumes:
      - application/json
      description: Removes a notification subscription for a group, tutor, or auditory.
      parameters:
      - description: Subscription identity
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/api.UnsubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unsubscribe from schedule notifications
      tags:
      - Notifications
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Admin-Key
    type: apiKey
swagger: "2.0"